\name{SpBuffer-package}
\alias{SpBuffer-package}
\alias{SpBuffer}
\docType{package}
\title{
\packageTitle{SpBuffer}
}
\description{
\packageDescription{SpBuffer}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SpBuffer}
\packageIndices{SpBuffer}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SpBuffer}

Maintainer: \packageMaintainer{SpBuffer}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
