\name{EpiBuffer-package}
\alias{EpiBuffer-package}
\alias{EpiBuffer}
\docType{package}
\title{
\packageTitle{EpiBuffer}
}
\description{
\packageDescription{EpiBuffer}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{EpiBuffer}
\packageIndices{EpiBuffer}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{EpiBuffer}

Maintainer: \packageMaintainer{EpiBuffer}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
